refactor(analysis-types): extract effort driver DTO
